tied firestore to central state. firestore transaction refactoring

+++ src/lib/db.ts
@@ -1,5 +1,5 @@
 import firebase from 'firebase/app'
-import { Album } from 'types'
+import { Album, Stat } from 'types'
 import { getUser } from './auth'
 import _ from 'lodash'
 
@@ -28,11 +28,11 @@ export const addPhotoToAlbum = async (albumTitle: string, photo: string) =>
         const albumDoc = await transaction.get(albumRef)
         const array: string[] = albumDoc.data()!.photos
 
+        if (array.includes(photo)) throw Error('Photo exists')
+
         const statRef = userRef.collection('stat').doc(photo)
         const statDoc = await transaction.get(statRef)
         const count: number = statDoc.data()?.count ?? 0
-
-        if (array.includes(photo)) return count
 
         transaction.update(albumRef, { photos: [...array, photo] })
             .set(statRef, { count: count + 1 })
@@ -43,14 +43,16 @@ export const addPhotoToAlbum = async (albumTitle: string, photo: string) =>
 export const removePhotoFromAlbum = async (albumTitle: string, photo: string) =>
     db.runTransaction(async transaction => {
         const userRef = getUserRef()
-        const statRef = userRef.collection('stat').doc(photo)
+
         const albumRef = userRef.collection('albums').doc(albumTitle)
-
-        const statDoc = await transaction.get(statRef)
         const albumDoc = await transaction.get(albumRef)
-
         const array: string[] = albumDoc.data()!.photos
-        const count: number = statDoc.data()!.count
+
+        if (!array.includes(photo)) throw Error('No photo')
+        
+        const statRef = userRef.collection('stat').doc(photo)
+        const statDoc = await transaction.get(statRef)
+        const count = statDoc.data()!.count as number
 
         transaction.update(albumRef, { photos: array.filter(e => e !== photo) })
 
@@ -65,16 +67,16 @@ type Rec = Record<string, number>
 export const deleteAlbum = async (title: string) => {
     const userRef = getUserRef()
     const albumRef = userRef.collection('albums').doc(title)
-    const statRef = userRef.collection('stat')
-
     const albumDoc = await albumRef.get()
+    if (!albumDoc) throw Error('No album')
+    
     const { photos } = albumDoc.data() as Album
-
     if (!photos.length) {
         albumRef.delete()
         return [] as string[]
     }
-
+    
+    const statRef = userRef.collection('stat')
     const statDocs = await statRef
         .where(firebase.firestore.FieldPath.documentId(), 'in', photos)
         .get()
